refactor: spark 관련 로직 제거
