+++ src/main/java/chess/dao/BoardDaoImpl.java
@@ -1,17 +1,23 @@
 package chess.dao;
 
 import chess.domain.Color;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
 
-public class BoardDaoImpl implements BoardDao {
+@Repository
+public class BoardDaoImpl implements BoardDao{
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public BoardDaoImpl(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
     @Override
     public void save(Color turn) {
         final String sql = decideSql();
 
-        executeSave(turn, sql);
+        jdbcTemplate.update(sql, turn.name());
     }
 
     private String decideSql() {
@@ -22,56 +28,24 @@ public class BoardDaoImpl implements BoardDao {
     }
 
     private boolean existBoard() {
-        final String sql = "select id from board where id = 1";
-        try (final Connection connection = DBConnector.getConnection();
-             final PreparedStatement statement = connection.prepareStatement(sql)) {
-
-            final ResultSet resultSet = statement.executeQuery();
-            return resultSet.next();
-        } catch (SQLException e) {
-            throw new IllegalArgumentException("예상치 못한 에러가 발생했습니다. 다시 시도해주세요.");
-        }
+        final String sql = "select count(*) from board where id = 1";
+        final Integer integer = jdbcTemplate.queryForObject(sql, Integer.class);
+        return !integer.equals(0);
     }
 
-    private void executeSave(Color turn, String sql) {
-        try (final Connection connection = DBConnector.getConnection();
-             final PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, turn.name());
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new IllegalArgumentException("예상치 못한 에러가 발생했습니다. 다시 시도해주세요.");
-        }
-    }
 
     @Override
     public Color findTurn() {
         final String sql = "select turn from board";
-        return executeFindTurn(sql);
-    }
+        final String turn = jdbcTemplate.queryForObject(sql, String.class);
 
-    private Color executeFindTurn(String sql) {
-        try (final Connection connection = DBConnector.getConnection();
-             final PreparedStatement statement = connection.prepareStatement(sql)) {
-
-            final ResultSet resultSet = statement.executeQuery();
-            if (!resultSet.next()) {
-                throw new IllegalStateException("예상치 못한 에러가 발생했습니다. 다시 시도해주세요.");
-            }
-            return Color.from(resultSet.getString("turn"));
-        } catch (SQLException e) {
-            throw new IllegalArgumentException("예상치 못한 에러가 발생했습니다. 다시 시도해주세요.");
-        }
+        return Color.from(turn);
     }
 
     @Override
     public void deleteBoard() {
         final String sql = "delete from board where id = 1";
-        try (final Connection connection = DBConnector.getConnection();
-             final PreparedStatement statement = connection.prepareStatement(sql)) {
 
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            throw new IllegalArgumentException("예상치 못한 에러가 발생했습니다. 다시 시도해주세요.");
-        }
+        jdbcTemplate.update(sql);
     }
 }
